Switched to OpenMP.
64-bit version does not link. Something about virtual void I think.

git-svn-id: svn://136.177.114.72/svn_GW/IPhreeqc/trunk@6889 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ IPhreeqcLib.cpp
@@ -907,7 +907,7 @@ IPhreeqcLib::CreateIPhreeqc(void)
 		{
 			IPhreeqcPtr = new IPhreeqc;
 		}
-		n = IPhreeqcPtr->Index;
+		n = (int) IPhreeqcPtr->Index;
 	}
 	catch(...)
 	{
